return size_t where appropriate, cast to Uint32/Sint32 where necessary

+++ src/Camera.h
@@ -103,7 +103,7 @@ public:
 
 	// lights with properties in camera space
 	const std::vector<LightSource> &GetLightSources() const { return m_lightSources; }
-	const int GetNumLightSources() const { return m_lightSources.size(); }
+	const size_t GetNumLightSources() const { return m_lightSources.size(); }
 
 private:
 	RefCountedPtr<CameraContext> m_context;
