+++ src/scenegraph/Model.h
@@ -121,9 +121,9 @@ public:
 	//materials used in the nodes should be accessible from here for convenience
 	RefCountedPtr<Graphics::Material> GetMaterialByName(const std::string &name) const;
 	RefCountedPtr<Graphics::Material> GetMaterialByIndex(const int) const;
-	unsigned int GetNumMaterials() const { return m_materials.size(); }
+	size_t GetNumMaterials() const { return m_materials.size(); }
 
-	unsigned int GetNumTags() const { return m_tags.size(); }
+	size_t GetNumTags() const { return m_tags.size(); }
 	MatrixTransform *const GetTagByIndex(unsigned int index) const;
 	MatrixTransform *const FindTagByName(const std::string &name) const;
 	typedef std::vector<MatrixTransform *> TVecMT;
@@ -131,7 +131,7 @@ public:
 	void AddTag(const std::string &name, MatrixTransform *node);
 
 	const PatternContainer &GetPatterns() const { return m_patterns; }
-	unsigned int GetNumPatterns() const { return m_patterns.size(); }
+	size_t GetNumPatterns() const { return m_patterns.size(); }
 	void SetPattern(unsigned int index);
 	unsigned int GetPattern() const { return m_curPatternIndex; }
 	void SetColors(const std::vector<Color> &colors);
